通用属性修改-add-将属性类型改成double了

+++ Assets/Script/Game/Module/DataCenter/Components/ItemData.cs
@@ -11,7 +11,7 @@ namespace SGame
 		public struct Value
 		{
 			public int      id;
-			public long     num;
+			public double     num;
 			public ItemType type;
 
 			public override string ToString()
